+++ toolbox/+igraph/centrality.m
@@ -1,4 +1,4 @@
-function values = centrality(graph, method, graphOpts, methodOpts)
+function values = centrality(graph, method, graphOpts, methodOpts, attribute)
 %CENTRALITY calculate centrality measure in a graph
 %   VALUES = CENTRALITY(GRAPH, METHOD) calculate the centrality for all nodes
 %      in GRAPH using METHOD. METHOD can be one of 'closeness', 'harmonic',
@@ -7,6 +7,11 @@ function values = centrality(graph, method, graphOpts, methodOpts)
 %      nodes in the graph to calculate centrality on.
 %   VALUES = CENTRALITY(GRAPH, METHOD, ..., 'directed', TF) if true, treat the
 %      graph is directed (defaults to IGRAPH.ISDIRECTED).
+%
+%   GRAPH = CENTRALITY(GRAPH, METHOD, RESULT, "NAME") if GRAPH can hold
+%   attributes, setting RESULT will determine the name of a node attribute to
+%   store the results such that GRAPH.Nodes.NAME will be a vector of centrality
+%   values. Note: The original GRAPH is not modified.
 %
 %   Method specific arguments:
 %      'normalized'   logical for 'closeness', 'harmonic'; whether to return
@@ -34,6 +39,8 @@ function values = centrality(graph, method, graphOpts, methodOpts)
        methodOpts.normalized (1, 1) logical = true;
        methodOpts.damping (1, 1) ...
            {mustBeInRange(methodOpts.damping, 0, 1)} = 0.85;
+       attribute.results (1, :) char ...
+           {igutils.mustHoldNodeAttr(graph, attribute.results)} = '';
     end
 
     method = lower(method);
@@ -41,6 +48,18 @@ function values = centrality(graph, method, graphOpts, methodOpts)
         method = 'burt';
     end
 
+    if ~isempty(attribute.results) && ...
+            length(methodOpts.vids) < igraph.numnodes(graph)
+        error("Cannot store results in graph as node attributes " + ...
+              "when calculating centrality for only a subset of " + ...
+              "nodes (i.e. when vids is not a vector of all node ids.)");
+    end
+
     values = mexIgraphDispatcher(mfilename(), graph, method, graphOpts, ...
                                  methodOpts);
+
+    if ~isempty(attribute.results)
+        graph.Nodes.(attribute.results) = values';
+        values = graph;
+    end
 end