+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A main igraph help entry used to describe common function arguments.
 - Reindex membership.
 - Allow `multiple` option for algorithms that allow when using MATLAB graphs.
+- Options for storing results as node and edge attributes and using stored attributes in other functions.
 
 ### Changed